requirements.txt
google-cloud-bigquery==3.25.0
pandas==2.2.2
db-dtypes==1.2.0
tqdm==4.66.5
torch==2.4.0
torchvision==0.19.0
torch_geometric==2.5.3
sentence-transformers==3.0.1